Refactor out the getRegCount method into a seperate InvokeInstruction interface

+++ dexlib/src/main/java/org/jf/dexlib/Code/Format/Instruction35c.java
@@ -105,7 +105,7 @@ public class Instruction35c extends InstructionWithReference implements FiveRegi
         return Format.Format35c;
     }
 
-    public byte getRegCount() {
+    public short getRegCount() {
         return regCount;
     }
